update: recruitButton 위치 AnimatedTextSection으로  변경
